requirements.txt
# Email Automation Dependencies

# Load environment variables from .env file
python-dotenv==1.0.0

# Email handling is built into Python's standard library (smtplib, email)
# No additional packages needed for basic functionality

# Optional: For advanced HTML email templates
# jinja2==3.1.2

# Optional: For scheduling tasks
# schedule==1.2.0

# Optional: For better email validation
# email-validator==2.1.0

# Optional: For sending emails through different providers
# sendgrid==6.11.0
# boto3==1.34.0  # For AWS SES
